images: caching, thumbnails
The caching problem is resolved an the rln root-link is removed.
The _show_ files may as well be used for thumbnails so the _mini_
files are superfluous. The removing of them must be done with care,
since the server is also deeply involved. So far they are still
around, but 'mini' is replaced with 'show' in the <template> of
view-main.gjs, e.g. '<img src="{{item.show}}" ...' etc.