upload images
Updated node_modules with pnpm. Fixed some GUI issues etc. Started work with the
'upload new images' functionality from within the album-tool point of view with
some progress, but much more to do. Here the file names should be checked and
maybe modified if too trivial? It IS possible to check names against the sqlite
filename table. Is it desirable to do that? Or should the same name be possible
in different albums? If impossible, the dupname check is superfluous?! On the
other hand, an album's images may be entered in other ways, so dupcheck needed!