gui etc.
Changed some deprecated Ember uses and some GUI details. Working on the
image upload functionality.

ALSO fixed a detail that prevented automatic inclusion of new albums
with the adding of a  .imdb  file. It remains to fix that some thumbnails
are missing until a mouse click is done (almost  unimportant, though).