orphans
Eliminating orphan _mimi_ and _show_ files from each album that is put up in the
album tree of the main menu. Important, since the image count is based on the
counting of '_mini_... .png' images in the album. This will probably eliminate all
number discrepancies between 'count by find "/"' and the 'autocount from tree'.